1.15.2
Former-commit-id: 6f57f8404c3dd597626e8d37d1aff2af895a02d1

+++ packages/_helpers/fake_resources.js
@@ -20,37 +20,55 @@ const langResources = [
   // no Spanish linguistic resources
 ];
 
-function getPlaceHolder(libName) {
-  return `{"${libName}_PLACEHOLDER":"true"}`;
+function getPlaceHolder(rollup, libName) {
+  if (rollup) {
+    return `{"${libName}_PLACEHOLDER":"true"}`;
+  } else {
+    return '{}';
+  }
 }
 
-function fakeFiles() {
-  // rename
+function fakeFiles(rollup) {
   for (let i = 0; i < langResources.length; i++) {
     const name = langResources[i][1];
-    console.log(`renaming ${name}`);
-    fs.renameSync(name, `${name}.BU`);
-  }
 
-  // fake ones but with placeholder
-  for (let i = 0; i < langResources.length; i++) {
-    const newContent = getPlaceHolder(langResources[i][0]);
-    console.log(`creating fake content for ${langResources[i][1]}`);
-    fs.writeFileSync(langResources[i][1], newContent, 'utf8');
+    if (fs.existsSync(`${name}.BU`)) {
+      console.log(`backup file already exists: ${name}.BU => skipping`);
+    } else {
+      if (!fs.existsSync(name)) {
+        console.log(`original file does not exist: ${name} => skipping`);
+      } else {
+        // rename
+        console.log(`renaming ${name}`);
+        fs.renameSync(name, `${name}.BU`);
+
+        // fake ones with placeholder
+        const newContent = getPlaceHolder(rollup, langResources[i][0]);
+        console.log(`creating fake content for ${langResources[i][1]}`);
+        fs.writeFileSync(langResources[i][1], newContent, 'utf8');
+      }
+    }
   }
 }
 
 function cleanFakes() {
-  // delete fake ones
   for (let i = 0; i < langResources.length; i++) {
     const name = langResources[i][1];
-    console.log(`deleting ${name}`);
-    fs.unlinkSync(name);
-  }
 
-  // take from backup
-  for (let i = 0; i < langResources.length; i++) {
-    const name = langResources[i][1];
+    if (!fs.existsSync(`${name}.BU`)) {
+      console.log(`no backup file for ${name}!`);
+      continue;
+    }
+
+    if (fs.existsSync(name)) {
+      // delete fake ones
+      console.log(`deleting ${name}`);
+      fs.unlinkSync(name);
+    } else {
+      console.log(`fake file does not exist: ${name}`);
+    }
+
+    // take from backup
     console.log(`renaming ${name}`);
     fs.renameSync(`${name}.BU`, name);
   }
@@ -78,7 +96,7 @@ function replacePlaceholders() {
     const originalLib = fs.readFileSync(compFile, 'utf8');
     const targetValue = fs.readFileSync(resource, 'utf8');
 
-    const placeholder = getPlaceHolder(libName);
+    const placeholder = getPlaceHolder(true, libName);
     if (originalLib.indexOf(placeholder) == -1) {
       throw `cannot find ${placeholder} in ${compFile}`;
     } else {
@@ -88,13 +106,19 @@ function replacePlaceholders() {
   }
 }
 
-if (process.argv.indexOf('before') > -1) {
-  console.log('BEFORE');
-  fakeFiles();
-} else if (process.argv.indexOf('after') > -1) {
-  console.log('AFTER');
+if (process.argv.indexOf('before_tsc') > -1) {
+  console.log('before_tsc');
+  fakeFiles(false);
+} else if (process.argv.indexOf('before_rollup') > -1) {
+  console.log('before_rollup');
+  fakeFiles(true);
+} else if (process.argv.indexOf('after_tsc') > -1) {
+  console.log('after_tsc');
+  cleanFakes();
+} else if (process.argv.indexOf('after_rollup') > -1) {
+  console.log('after_rollup');
   cleanFakes();
   replacePlaceholders();
 } else {
-  throw 'error, put before or after';
+  throw 'error, wrong arg';
 }
